requirements.txt
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
Pillow>=9.0.0
streamlit>=1.25.0
numpy>=1.23.0
